fix manpage copy for osx

+++ recipes/ococo/build.sh
@@ -26,5 +26,7 @@ export LIBRARY_PATH=$LIBRARY_PATH:${PREFIX}/lib
 make VERBOSE=1
 mkdir -p ${PREFIX}/bin
 cp ococo ${PREFIX}/bin
-cp ococo.1 ${PREFIX}/share/man/man1
+if [ -e ${PREFIX}/share/man/man1 ]; then
+    cp ococo.1 ${PREFIX}/share/man/man1
+fi
